+++ autogen/src/main.rs
@@ -8,8 +8,7 @@ use anyhow::{anyhow, Result};
 use gen_gql_schema::update_gql;
 use schemars::schema::RootSchema;
 use serde_json::{json, Value};
-use tailcall::cli::init_runtime;
-use tailcall::config;
+use tailcall::{cli, config};
 
 static JSON_SCHEMA_FILE: &str = "../generated/.tailcallrc.schema.json";
 
@@ -47,7 +46,7 @@ async fn main() {
 
 async fn mode_check() -> Result<()> {
     let json_schema = get_file_path();
-    let rt = init_runtime(&Default::default(), None);
+    let rt = cli::runtime::init(&Default::default(), None);
     let file_io = rt.file;
     let content = file_io
         .read(
@@ -73,7 +72,7 @@ async fn mode_fix() -> Result<()> {
 async fn update_json() -> Result<()> {
     let path = get_file_path();
     let schema = serde_json::to_string_pretty(&get_updated_json().await?)?;
-    let rt = init_runtime(&Default::default(), None);
+    let rt = cli::runtime::init(&Default::default(), None);
     let file_io = rt.file;
     log::info!("Updating JSON Schema: {}", path.to_str().unwrap());
     file_io
