+++ src/cli/tc.rs
@@ -12,7 +12,7 @@ use super::update_checker;
 use crate::blueprint::{validate_operations, Blueprint, OperationQuery, Upstream};
 use crate::cli::fmt::Fmt;
 use crate::cli::server::Server;
-use crate::cli::{init_runtime, CLIError};
+use crate::cli::{self, CLIError};
 use crate::config::reader::ConfigReader;
 use crate::config::Config;
 use crate::print_schema;
@@ -25,7 +25,7 @@ pub async fn run() -> Result<()> {
     let cli = Cli::parse();
     logger_init();
     update_checker::check_for_update().await;
-    let runtime = init_runtime(&Upstream::default(), None);
+    let runtime = cli::runtime::init(&Upstream::default(), None);
     let config_reader = ConfigReader::init(runtime.clone());
     match cli.command {
         Command::Start { file_paths } => {
