+++ src/lib.rs
@@ -21,7 +21,7 @@ pub mod lambda;
 pub mod mustache;
 pub mod path;
 pub mod print_schema;
-pub mod target_runtime;
+pub mod runtime;
 pub mod try_fold;
 pub mod valid;
